Wellesley fresh app - now maybe with hours
I'm in the process of adding the hours in to the Wellesley Fresh app. I
think that this will be really great information to have tbh.

Although I'm also not so sure how it will work.

I also finally got myself updated-ish to XCode 7.3.1. I really want to
get Xcode8 but the wifi here sucks so one step at a time.

I just hope that nothing breaks when I move up to 8.

I had to make a lot of classes thta would abstract information away from
me, such as the DayOptions class, the DiningHours class, and the
DiningHall class.