Added hours for all dining halls
I finally got the hours added in for all of the dining halls, plus I
added the UIProgressViews that show how much time has elapsed so far. I
think it looks pretty good - the only problem is that some of the labels
have to be extra small. I think I'll fix that ata some point.

I even have it show how much time is left until it's open again for the
next meal. I think that this would be really helpful for all of the
rtail centers as well, although I will probably have to make some kind
of tableview tableviewcell so that I can have multiple header cells
within the table.