Shows nearest menus now, refactored a little
I wanted to have the nearest menus available so I finally put in some
code to do that. Unfortunately though I had to copy and paste it from
the SecondViewController which isn't the best solution, although I'm not
sure what would be a better solution. I guess I could have an object
that comprises these methods but where you have to pass in the
objects...but that seems like it could almost get overly complicated.

I also got rid of the refresh control on the second view controller
because it technically wasn't doing anything by the time I commented out
all of its code. Eevntually I'll just delete the old code.

I want to really get this code refactored though because it's kind of
gross...and way too long...and I don't like it haha.

All I did was rearrange things so that they made a little bit more
sense.

I also ended up getting rid of a lot of the warnings by updating the
code to more Swift 2/3 standards and not doing things that are going to
be deprecated/are deprecated.