Fixed HourRange, fiddled with layout
HourRange was giving a negative answer for time left which was
just...not right. Something is still kind of "up" with that part of the
code that I really need to look into.

I also fiddled with the layout of the firstviewcontroller, more
specifically, the tableview. I wanted to have it take up all of the
space remaining after the label because there was a large gap on the 6s
Plus so I just added a little bit more to the viewWillAppear code to
programmatically determine how tall and where the tableview will be.