Fixed hour ranges and dining hall things
The Hours of Operation page controlled by HoursViewController is now
completely correct. There's still some strange things going on with the
hours of operation when it changes from lunch to dinner or something
like that but I can only test those cases at certain times of the day.
Shouldn't there be a way to change what time it is and therefore be able
to test it more effectively?

Well anyway, it works now. I had to change a lot of the logic in order
to do so, and I also made the hour range much more concise. I'll delete
commented-out code soon. I had to change logic in dining hall class
because otherwise it would give inbetween even if it wasn't the right
interval.