Fixed DiningHall and HourRange
DiningHall was giving the wrong hour range to get percent elapsed from,
so I fixed that and made the code, again, more concise.

I also had to fix a little bit with how withinRange is determined.
Because I had both of them non-inclusive, in between the times becomes
ambiguous, so it would do something strange. I seem to have fixed it
though, and was able to test it.