Added CustomButtonCell Class, Added to More Page
Now the more page also includes everything that I needed for giving
feedback to the dining halls, and a section for settings, etc. I might
want to have a specific setting that allows you to rank dining halls by
whatever favorite, so that it will preload that one when you open the
AllDiningHalls page.

I also want to fix the problem of the UIPicker on Nearby/first view
controller not having anything when you first open it, which would be a
problem.

I also want to fix the problem of the UIPicker on AllDiningHalls/Second
View Controller being behind the lines of the table view.

I might also change ThirdViewController to a navigation controller + a
tableviewcontroller.

I'm using the CustomButtonCell class for notifying different dining
halls. It's a special TableViewCell that only has a button inside it.