Fixed problems with tableviewcontroller
SecondViewController, which inherits from UITableViewController, was not
showing the headers' names for some reason whenever you picked a new
menu to show. I also could not for the life of me figure out what the
UIPickerView was showing up behind the table view. I just had to add it
to the navigation controller's view though, go figure.

Also fixed a problem with the headers. The grouped headers seem to be
overlapping the titles of the sections, so I just had it get rid of
its own label and made a different class for just the headers and the
grouped headers. I also made it so the text of the header is closer to
the cells it pertains to.