Fleshed out DayHourRange, Changed HourRange
I added DayHourRange-specific functions that either override or
elaborate on the HourRange ones.

I also created a system in HourRange such that I can override certain
functions in DayHourRange but other functions such as percentTimeElapsed
don't have to change at all because they'll just call the same functions
that are in the DayHourRange class instead of the HourRange class.

I have not implemented that part in the DayHourRange class yet but I
just wanted to commit before anything bad happens.

I'm pretty sure there's some kind of errors in another file so I'll need
to fix that as well.