Completed (maybe) the DayHourRange class
I still have to implement differences in months and such, it really
should not be any more than one month's difference if anything, and even
if they are in different years that shouldn't matter because all of the
months always have the same days except for february, which isn't at
some kind of year boundary.

The map has begun jumping around again so I'm going to try to remedy
that.