Hour Range should totally work, Added disclosure
I added in the little ">" disclosure indicator to the About button,
which is nice to know for future reference.

I also made hour range basically foolproof, although I'm going to have
to test that a lot. I might use unit tests for this purpose.

Things are looking fairly good now, which is nice.