Fixed missing header on table view, updated readme
For some reason, the header cell was missing on the table view on first
view controller, so I finally added it back on. Apparently you need to
have the header height returned in a delegate function.

I also updated the readme so that it is current.

Additionally, I just thought of something that was really missing:
bagged lunch form and the ability to see future menus.