Added more features, Added more retail centers
We now have the Leaky Beaker, the Emporium, and Collin's Cafe.

Unfortunately, it seems to be crashing a bit - at least until I stopped
forcibly unwrapping some values. I'm not sure if this fixed it
completely though.

I'm going to put it on my phone to see.

I'm also going to need to debug retail centers' hours' behavior a little
further just to be sure.