Added self., deleted extraneous tags
Apparently some kit service freaks out if you don't have self. in front
of all of the fields that are in your class, so I'm slowly adding more
and more of these.

I also deleted the extraneous @IBAction tags from the
ThirdViewController because I also deleted the old buttons from the
storyboard.

I've tested this, and it seems fine with it.