Fixed potential bugs with DayHourRange
Changed the day formatter to use 365-days instead of the days of the
month.

(Maybe everything would have been easier if I'd just used the number of
days since 1970, the absolute time reference for everything, but you
know)

I also used leap years in order to calculate what a high date that is in
the next year should be added to in order to not make the a negative
number just because it's at the end of December.

Of course, that functionality is useless because we're not at Wellesley
that late in December, but it makes things complete at least if I do
something like this in the future.