Fixed a couple of bugs
Progress cells now seem to stay as they are...at least for now. I might
just "roll my own" version later on if I get frustrated enough with it.

I also found out that I was only returning hours from the DayHourRange,
which explains why it was so far off. So now it's return the minutes and
seconds added in as well.