Tried fixing progressview's appearance
For some reason, they refuse to show increments that are less than an
hour for Bates, Tower, and Pom. I'm very confused.

I think this may have something to do with the fact that they're on an
DayHourRange instead of an HourRange, but it's also never done this
before. It literally does not show up until it gets to an hour.