Added place to store some assets, closes #35, #36
I added a folder to store some in-progress assets such as photoshop
files. I normally would put these in some other remote corner of my
laptop but I really don't want to have to run around and find them this
time around.

I also added a Contact button in ThirdViewController that is immediately
connected to a mail view controller, and even put in the version number
and such so that I can filter emails for this, which should be great.