Fixes #34, bug with feet
The reason why it couldn't find the proper number of feet was because my
limit was too small, around 100,000 or so. Anything larger than that
would not be made a minimum because it's inherently too large. So I
increased that by several powers of ten and it's good to go now.

I'm still working on the other bug - the bug with the location showing
up in stone d. I'm not sure why this is. I should probably test it on
other phones.