Continuing to work on #60
I can change the pin colors now, but the pin colors have no bearing on
which dining hall they're for. In fact, it seems sort of inverse in a
way - or perhaps it's that the location changes all of a sudden? I just
found my own bug while writing this commit message.