Fixed #67: got rid of extraneous lines
Table view cells have lines on the bottom that don't go away if you're
the last cell in the section and there are no sections after you, so I
had to add header cells to the very end of the table views in order to
make this go away.

Also, I changed some of the constraints on the about page.