Fixed #79: hours were too much
By too much, I mean that the number of hours that Collins Cafe and the
Leaky Beaker were purportedly going to be closed by was as much as a
year.

As in they would be closed by a year.

It all turned out to be some logic mistakes in DayHourRange for
getHigherDay and getHighDay.