Fixes #81: turn off sticky headers
I only wanted to do this for settings because basically you don't need
to have such a thing for settings

All you had to do was change the style from .plain to .grouped. Because
I wasn't using a tableviewcontroller, and just instantiating a
tableview, I just had to use the tableview init method where you give it
a frame and a style.