Fixes #75: rename view controllers
Besides HoursViewController, all of them had very uninspired names, so I
just wanted all of that to be fixed, so I changed them from First,
Second, and ThirdViewController to Nearby, Menu, and
SettingsViewController, respectively.

This was because I kept getting confused and really had to think about
what was happening a lot of the time.