Accidentally Created #92
Issue #92 - fix constraints for about page

So basically I broke the about page because I was trying to be too fancy
with all of my views.
Gosh dangit.

I should have just programmatically done this a long time ago.

I can even have an enum for the different indents and stuff like that.