Fixed #92, Closes #91
Issue #92 - fix constraints for about page
I judged the about page as it was to be unsalvageable so I just redid
the whole thing but using a table view with cells that could be set to
different styles. I still need to make the buttons all fit their text
though.

Issue #91 - alert user if their iphone is in airplane mode lol
I think this issue may have been corrected but it still needs to be
tested on an actual iPhone, and I've been a bit lazy lately.
I did however add the necessary information to the plist for this to
happen, so it should be fine, because that's what the internet said all
there was to it.