Fixed #95, Fixes #96
Issue #95 - actually alert user about airplane mode
The problem was the app was crashing if you opened it with airplane mode
on, so it couldn't tell you to go to settings and turn off airplane
mode.

Issue #96 - fix regexes
The regexes for getting rid of the junk from Wellesley Fresh weren't
doing so well so I fixed that as well, and likely will have to simplify
my regexes in the future.