Fixed #97 - stop selection of about page stuff
Because I made the about page out of a table view, it only made sense
for me to stop it from selecting the cells because otherwise, it would
highlight a given cell in gray, which wasn't very aesthetically
pleasing.

So I changed the selection style of the cells to none.