Worked on #99 - cont refactoring newCellsInsertion
Added a function that takes care of making bottomHalfIndexPaths, the
second half of all of the index paths, and also a function that takes
care of updating the table view.

Once I delete the commented out code that these two functions replace,
the whole file should be at least a little shorter.

Also had to make an absolute difference function so that I could figure
out, without having to use too much logic, what the absolute difference
was between the original array and the new array.