Resolved #99 - refactor code for MenuVC
Refactored two functions that really needed cleaning up - load and
newCellsInsertion, not to mention deleted a ton of old code that either
did not work or wasn't even used anywhere. There are a ton of new helper
functions in here as well. I deleted all of the commented out code that
I generated in the process of refactoring as well.

Also commented out the debugging line that makes the app download a full
day's menu, which I use for testing purposes, so this code is pretty
much in its final form for deployment.