Fixed #100 - fix height of tableview
This specifically meant the one in MenuViewController. It was sometimes
too long, and sometimes too short. It was because I was manually
adjusting the height of the table view instead of letting it do that
itself, after all, it does automatically determine the height of the
cells. Once I commented out the code that was calculating the height, it
started working perfectly.