Started looking into unit testing
Working on Issue #66 - write unit tests
I really should split this issue into far smaller issues because I
really do not think that it is really feasible or practical for me to
write unit tests for all 10+ classes and have it be under the umbrella
of a single issue. Each class in itself should probably have its own
issue.
Anyway, Swift doesn't seem to have a good mocking framework like
AngularJS does, with Jasmine, but that's probably because of the newness
of Swift and the unpopularity of unit testing among iOS developers,
which I find strange.
It is rather difficult though for me to test the code because I barely
have any functions that take arguments, which is a problem. I also
haven't really structured the code to be friendly for dependency
injection, like I did in Angular, so I'm going to have to write some
workarounds for that.
I'm really not sure either if I should be testing the view controllers
as they are now because the code is a lot of just making alterations to
various instance variables of the view controller...well, I guess I can
do something about that.