Fixed #101, Fixed #102
Issue #101 - download content during first view
I, somewhat needlessly, had MenuVC use a new class, DataLoader, to load
its data for it - it just packages up some nice little things like
methods that check for the existence of previous days and such - and
made a method for MenuVC that downloads the content. No matter where you
download the content though, it all goes to the same place, but I guess
now it just does it in a little more orderly fashion?
Anyway, in NearbyVC, it checks for the existence of one more view
controller in the navigation controller, then picks the second one (at
index 1), and gets its top view controller (aka MenuVC) and makes it run
the method loadData so that the data is already all loaded by the time
you open up the app.
I know this works because I set DataLoader to think that it needs to get
data every single time it loads up because I made the todayString and
the value in the UserDefs different. However, I changed it back before I
committed this.

Issue #101 - fix sizing of tableview in nearbyVC
I hadn't tested it in a while, so I ended up putting some longer than
one line data in there, and as it turns out, the table view is too
long. I could seet that because I couldn't see the round end of the
scroll indicator if I scrolled all the way to the bottom. Anyway, it
turns out that it was the height of the tab bar too long for the screen,
so once I subtracted the height of the tab bar, it fits perfectly. I
originally thought that this was some sort of issue with the tableview's
autoresizing.