Resolved #103, Working on #21
I kind of want to put off preparing for submitting the app, so I decided
to work on clearing some of the backlog of issues.

Issue #103 - add ability to turn on/off scrolling to next meal
Added the setting in SettingsViewController. Had to create a new cell
class for it too. I needed a cell that would have a switch inside it so
that you could turn it on and off. In SettingsVC, I also added two
arrays: switchCells and switchSelectors. switchCells has the names of
the switchCells so that you can tell what a cell is supposed to be, a
regular cell, a switch cell, etc. switchSelectors just helps me to not
have to hardcode some kind of switch statement into the already bloated
cellForRowAt method.

Issue #21 - first view - show the meal that is going to be served next
Working on this meant that I had to have some kind of convenient method
for finding the meal that is either being currently served or the meal
that will be served soon. I have been implementing new methods in
HourRange and DiningHall to this end, although I'm not done with that
yet.
This issue also goes to hopefully the end goal of downloading tomorrow's
data as well, so that you can see what's going to be for dinner
tomorrow. I haven't gotten that far ahead of myself yet. I just think
that this specific feature should be really useful for users - after
all, you don't really pull up the menu at lunch in order to see what the
choices were for breakfast, you're planning for the future.