Patching up #103, Working on #21
I forgot to change the isOn value of the switch to the value stored, so
that was added in this commit to the SettingsVC.

I also finished up the code in DiningHall/DiningHours/HourRange that
finds either the next meal or the current meal.

I implemented checking for and assigning a boolean value for
tableScrollsToNextMealKey in NearbyVC.

This is all in preparation of the next phase: actually having the table
scroll.