Resolved #21 - first view - show meal served next
Issue #21 - first view - show the meal that is going to be served next
At the moment, this only works with NearbyViewController, but I'm going
to implement it soon with MenuViewController. I think I should also
highlight the row that is the beginning of that section, at least until
I actually split up the menus by meal, but until then, I think that this
is good enough.
I currently have it not animating the scrolling, because it seemed to
freak out a little when I did that, probably because it's already
animating so many other things, such as cell insertion.