WellesleyFresh changed a bunch of stuff
It turns out that WellesleyFresh changed the format of their menus such
that they're storing things as jsons now - which could be more
convenient for me, but I also have no idea how things are being setup.
Like there doesn't seem to be a late night option? Very confusing. I'm
not sure why they decided to do this...but I might just submit the app
anyway, and work on getting a server up that can handle the getting of
the data for me...or something like that...