Tried to adapt to new wellesley fresh stuff
It's a real piece of work...JSON in Swift is a really unwieldy because
of the partial/dynamic-ish? typing. With JavaScript, JSON is a breeze
because Javascript barely has types as it is, which is amazing when
you're working with such a mixed bag of things...but if it's anything
even remotely typed, all of the conversions start murkying the waters.