Still refining DataLoader
I'm not sure how much longer they'll be using the old system, but it
seems like I have to use both systems in order to make this work, which
is just very odd and kind of crappy.

Anyway, apparently I had the BPLC's name wrong the entire time, so I
fixed that in Nearby and Menu VCs.

I want to have the tab bar controller store things like the names of the
dining halls and such so that I don't have to retrieve a copy from user
defaults, I can just get the same copy. I did this in cube timer, which
was really a quite nice way of organizing things.