Fixed #106, Probably Fixed #109
Issue #106 - save data from json somewhere
The problem with getting this finally all together was I didn't seem to
be able to get the dining halls straight, so I was giving the old
version of the dining halls to new dining hall functions. I think I
might have to make an entirely different class just for the new dining
hall functions.

Issue #109 - fix "smart download"
I believe I have this fixed - I have it save the data for the entire
week for every dining hall, such that you never have to download that
data twice. If it has the data, it uses it to make its menus. If it
doesn't, it won't use it, and if it does have the data but doesn't seem
to be able to parse it, it will get new data.