Fixed #114 - don't let nearby populate table
Issue #114 - don't let nearbyVC populate tableview if the data is old
This is a problem because I commented out some code in order to debug
things and I forgot to uncomment out the code once I was done with
debugging things.
So in this version, it is not commented out any longer.