Fixed #115 - put app icon at top of credits
I just thought that the app icon was a really major thing that needed to
be given more credit than the position at the bottom gave it? Not that
it's really in any particular order, I just thought the order was
actually really evidence of how things were added chronologically
instead of in any sort of meaningful way, and so I wanted to make it at
least a little more meaningful. Of course, I might have done this
earlier had I not had the worst possible way of structuring my about
page in the first place. Now that it's programmatically determined, I
can take more time to care about the layout.