Fixed #84, Fixed #85, Solved #113
Issue #84 - titles seem to overlap for header on iphone 7 plus
Issue #85 - See if I can replicate overlapping again some other time
Issue #113 - have menu uipickerview with default
It seems thta #84 and #85 have more in common with #113 than I
originally thought. It seems as though that because title being assigned
when you change dining hall with the pickerview, and the title being
assigned by the UITableViewController is going to a different label than
the nameLabel of the header class, the two overlap. When they're the
same, it's fine, but if they're different, it's horribly, horribly
wrong. And they would be different if it's telling it to populate the
label with, say, the shortened version of the dining hall's name rather
than the long version.
This commit takes care of all of these problems.