Solved #112 - have it auto populate TVs w menu
Issue #112 - have it automatically populate table views with menu
I technically only got it to do this for one table view - MenuVC - but I
think that this is the step to doing more with the DataLoader class.

Of course, I would like to have the DataLoader class under 500 lines,
but baby steps, baby steps.