Working on #116 - add sections to menus
I'm trying to get the menus to be split up into different sections so
that they seem better organized. I think unfortunately that this will
also end up changing how I make things in general though. Now that I
really think about it, I'm not sure how this would work with adding new
cells. The newCellsInsertion function is fragile enough as it is...