Closed #117, Incremented app version
I just realized that the app version is now out of date, so I updated
it.

Issue #117 - add the hall's name to the title of the menu view
controller
I think this is a small but important change, because I was thinking of
making the tableview into one with several sections.
Ideally, I would also like to fit the date somewhere into this whole
thing, but I'm not exactly sure if that would even fit.