Resolved #118 - make time fonts monospaced
Issue #118 - make fonts for times in progresscell monospaced
Whenever the time changed, especially going from a fatter digit like 0
to a thinner digit like 1, the size of the progressview would change, as
would the width of the label. It just doesn't look good like that,
changing all the time. So I thought I would make that difference.