Fixed #119 - time label constraints
Issue #119 - fix time constraints so that progressviews and such don't
move around
I think I got these fixed in general. I feel like the progressview is a
little bit too short but I think this can be altered soon. The best part
about this is that everything seems to be stable, finally.