Fixed #116: Fixed #129, Fixed #130
Issue #116 - add sections to menus
This was solved by the last commit as well as the following two issues.

Issue #129 - bold meals
Bolded the font of meals so that they would stand out better. I have it
done programmatically.

Issue #130 - fix separators for meals
I also turned off separators for not only the "meal" cells but for the
cells that immediately precede the meals, so that they wouldn't look too
weird. This should make things look better, without trying to change the
entire structure of the meals table view.