Fixed #134 - prevent crashing because weekends
Over the weekends, Stone Davis' dining hall closes and this really
throws the app for a loop. So I fixed the crashing by using the nil
coalescing operator, shorthand for (__ == nil) default : other thing.
This seems to have helped. I also prevented DiningHall from taking a
negative number as an index in one of its methods.