Fixed #139 - weekend hours bug
The weekend hours for StoneD weren't really right - in fact, they
weren't right for Collins Cafe or the Leaky Beaker either. So I had to
correct them. It seems like it was something to do with the way that I
was getting the lowHour/lowDay and saving it and such.