Continued to add to Menu, less warnings
I'm trying to get rid of the compiler warnings by making it happier and
commenting out or deleting redundant or pointless things.

I also continued to make Menu better. Although I think from this point
on, I'll have to branch off in order to be comfortable with really
messing around with things.